// src/app/api/model/aggregateCommandSet.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CommandEntry } from './commandEntry';


/**
 * A set of commands that you can display (for example on the Metasys User Interface) as a single command. The individual commands can be listed as parameter choices of the aggregate command.
 */
export interface AggregateCommandSet { 
    /**
     * The command identifier  The values for this property come from the `commandIdEnumSet` enumeration. Find the definition of `commandIdEnumSet` at the relative URL of `/schemas/enums/commandIdEnumSet`.  **See also:** * [Get enumeration](#operation/getEnumerationAsJsonSchema) for more information on retrieving an enumeration.
     */
    id: string;
    /**
     * A grouping of related commands.
     */
    commandSet: Array<CommandEntry>;
    /**
     * Always `true` for an aggregate command.
     */
    aggregateCommand: boolean;
    /**
     * Always missing or `null` for an aggregate command.
     */
    invokeUrl?: AggregateCommandSet.InvokeUrlEnum | null;
    /**
     * The translated name of this aggregate command.
     */
    title?: string;
}
export namespace AggregateCommandSet {
    export const InvokeUrlEnum = {
    } as const;
    export type InvokeUrlEnum = typeof InvokeUrlEnum[keyof typeof InvokeUrlEnum];
}