// src/app/api/model/putCommandRequest.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WritePriorityEnumSet } from './writePriorityEnumSet';


export interface PutCommandRequest { 
    priority?: WritePriorityEnumSet;
    parameters?: Array<any>;
    annotation?: string;
}
export namespace PutCommandRequest {
}